Move toolbar settings into dedicated ownership
