+++ browser/Reynard/Client/Preferences/BrowserPreferences.swift
@@ -35,6 +35,7 @@ final class BrowserPreferences {
     }
     
     func registerDefaults() {
+        migrateToolbarSettingsIfNeeded()
         let donationRecommendationShowTimeKey = key("HomepageSettings", "donationRecommendationShowTime")
         if UserDefaults.standard.object(forKey: donationRecommendationShowTimeKey) == nil {
             let delay = TimeInterval.random(in: (3 * 86_400)...(5 * 86_400))
@@ -88,10 +89,10 @@ final class BrowserPreferences {
             key("AppearanceSettings", "showsFullWebsiteAddress"): false,
             key("AppearanceSettings", "showsLandscapeTabBar"): true,
             key("AppearanceSettings", "defaultPageZoomLevel"): PageZoomLevels.defaultLevel,
-            key("AppearanceSettings", "bottomToolbarActions"): BottomToolbarAction.defaultActions.map(\.rawValue),
-            key("AppearanceSettings", "toolbarButtonHapticsEnabled"): true,
-            key("AppearanceSettings", "closeTabLongPressOpensNewTab"): true,
-            key("AppearanceSettings", "newTabLongPressClosesTab"): false,
+            key("ToolbarSettings", "bottomToolbarActions"): BottomToolbarAction.defaultActions.map(\.rawValue),
+            key("ToolbarSettings", "toolbarButtonHapticsEnabled"): true,
+            key("ToolbarSettings", "closeTabLongPressOpensNewTab"): true,
+            key("ToolbarSettings", "newTabLongPressClosesTab"): false,
             
             // Languages
             key("LanguageSettings", "websiteLanguages"): (try? JSONEncoder().encode(WebsiteLanguageCatalog.defaultLanguageCodes())) ?? Data(),
@@ -130,6 +131,23 @@ final class BrowserPreferences {
             ),
             forKey: Self.openLinksInAppsBridgeKey
         )
+    }
+
+    private func migrateToolbarSettingsIfNeeded() {
+        let defaults = UserDefaults.standard
+        for name in [
+            "bottomToolbarActions",
+            "toolbarButtonHapticsEnabled",
+            "closeTabLongPressOpensNewTab",
+            "newTabLongPressClosesTab",
+        ] {
+            let destinationKey = key("ToolbarSettings", name)
+            guard defaults.object(forKey: destinationKey) == nil,
+                  let existingValue = defaults.object(forKey: key("AppearanceSettings", name)) else {
+                continue
+            }
+            defaults.set(existingValue, forKey: destinationKey)
+        }
     }
     
     func bool(forSetting setting: String, key name: String) -> Bool {
@@ -711,10 +729,14 @@ final class BrowserPreferences {
             }
         }
 
+    }
+
+    // MARK: - Toolbar
+    struct ToolbarSettings {
         static var bottomToolbarActions: [BottomToolbarAction] {
             get {
                 let rawValues = prefs.stringArray(
-                    forSetting: "AppearanceSettings",
+                    forSetting: "ToolbarSettings",
                     key: "bottomToolbarActions"
                 ) ?? BottomToolbarAction.defaultActions.map(\.rawValue)
                 var seen = Set<BottomToolbarAction>()
@@ -723,14 +745,14 @@ final class BrowserPreferences {
                 }.filter {
                     seen.insert($0).inserted
                 }
-                return Array(actions.prefix(BottomToolbarAction.maximumVisibleActions))
+                return Array(actions.prefix(BottomToolbarLayoutPolicy.maximumConfiguredActions))
             }
             set {
                 var seen = Set<BottomToolbarAction>()
                 let actions = newValue.filter { seen.insert($0).inserted }
                 prefs.set(
-                    Array(actions.prefix(BottomToolbarAction.maximumVisibleActions)).map(\.rawValue),
-                    forSetting: "AppearanceSettings",
+                    Array(actions.prefix(BottomToolbarLayoutPolicy.maximumConfiguredActions)).map(\.rawValue),
+                    forSetting: "ToolbarSettings",
                     key: "bottomToolbarActions"
                 )
                 NotificationCenter.default.post(name: .bottomToolbarActionsDidChange, object: nil)
@@ -739,29 +761,29 @@ final class BrowserPreferences {
 
         static var toolbarButtonHapticsEnabled: Bool {
             get {
-                prefs.bool(forSetting: "AppearanceSettings", key: "toolbarButtonHapticsEnabled")
+                prefs.bool(forSetting: "ToolbarSettings", key: "toolbarButtonHapticsEnabled")
             }
             set {
-                prefs.set(newValue, forSetting: "AppearanceSettings", key: "toolbarButtonHapticsEnabled")
+                prefs.set(newValue, forSetting: "ToolbarSettings", key: "toolbarButtonHapticsEnabled")
             }
         }
 
         static var closeTabLongPressOpensNewTab: Bool {
             get {
-                prefs.bool(forSetting: "AppearanceSettings", key: "closeTabLongPressOpensNewTab")
+                prefs.bool(forSetting: "ToolbarSettings", key: "closeTabLongPressOpensNewTab")
             }
             set {
-                prefs.set(newValue, forSetting: "AppearanceSettings", key: "closeTabLongPressOpensNewTab")
+                prefs.set(newValue, forSetting: "ToolbarSettings", key: "closeTabLongPressOpensNewTab")
                 NotificationCenter.default.post(name: .bottomToolbarShortcutsDidChange, object: nil)
             }
         }
 
         static var newTabLongPressClosesTab: Bool {
             get {
-                prefs.bool(forSetting: "AppearanceSettings", key: "newTabLongPressClosesTab")
+                prefs.bool(forSetting: "ToolbarSettings", key: "newTabLongPressClosesTab")
             }
             set {
-                prefs.set(newValue, forSetting: "AppearanceSettings", key: "newTabLongPressClosesTab")
+                prefs.set(newValue, forSetting: "ToolbarSettings", key: "newTabLongPressClosesTab")
                 NotificationCenter.default.post(name: .bottomToolbarShortcutsDidChange, object: nil)
             }
         }
